Update hooks related to consumer secrets
